Reverted all changes in GridSearchHandler. Few renamings.
